fix printf format warnings

+++ PLC_esp8266/main/DataMigrations/Settings_20250413.h
@@ -98,7 +98,7 @@ namespace MigrateSettings {
             strcpy(pCurrSettings->datetime.sntp_server_secondary, "pool.ntp.org");
             strcpy(pCurrSettings->datetime.timezone, "Etc/GMT-3");
 
-            ESP_LOGI("Settings_20250413", "Migrate to %08X", DataMigrate.Version);
+            ESP_LOGI("Settings_20250413", "Migrate to %08X", (unsigned int)DataMigrate.Version);
         }
 
         inline void MigrateDown(void *pCurr, void *pPrev) {
